requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
sqlalchemy==2.0.36
pymysql==1.1.1
cryptography==44.0.0
alembic==1.14.1
redis==5.2.1
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
bcrypt==4.2.1
jinja2==3.1.5
httpx==0.28.1
stripe==11.4.1
resend==2.5.1
openai==1.59.6
apscheduler==3.11.0
svix==1.44.0
google-cloud-firestore==2.19.0
gspread==6.1.4
google-auth==2.37.0
